Created utility to manage position of comments.
Very naive approach. Meant to be used as follows.

    align = require('lib/align');
    align.move(index, $comments);
    align.above(index, $comments);
    align.below(index, $comments);

This algorithm is greedy in the sense that when a comment is
highlighted, it will appear next to the corresponding line of the code.
So rather than maximizing the number of visible comments on the screen,
we choose to display the comment as close as possible to the content it
refers to.

Unsure as to whether we can say that if we reach a comment that does
not need to move, then we can break out of the loop, i.e. while
adjusting the position of the comments above, upon reaching a comment
that does not need to be moved, would comments prior to that one still
need to be adjusted, or no?