Changed so that entire file is not rendered at once.
Added support to dynamically change the height of the editor with the
size of the window. The major accomplishment here is to handle the
scrolling of the editor with regard to the comments.

We do this by keeping track of the offset to where the editor has been
(vertically) scrolled, and apply a delta to the `top` css-property
relative to this.