Wrote handlers for when list of comments changed dynamically.
Adds a new comment box to the displayed DOM when a comment is added to
the list. Removes the comment box from the displayed DOM when it is
removed from the list.

Modified approach to not require binding a handler for each comment when
the cursor changes. Note that to handle the alignment of the comment
boxes when the window is resized and there are NO selected comments,
choose (arbitrarily) to align with respect to the first comment.