Fixed a couple style issues with displaying comments.
Changed so that the scolling is "smooth" for the comments, and that
they do not immediately disappear from sight when region falls below
what would fit the ENTIRE comment.

Also made it so that the width of the comment has the same upper bound,
regardless of whether it is selected (and thereby shifted left) or not.