Added UI interaction for user to edit comments.
Makes use of the `contenteditable` attribute that is supported by ALL
browsers. (pretty cool!)

When the user clicks the edit button, the <div> becomes editable. Since
the comment text may contain newlines, we convert these to <br> tags.

It may be nicer to have the footer hidden unless the comment is
selected, but then we need to deal with the changing sizes of the boxes
and the relative alignment.